requirements.txt
Django==4.2.7
google-generativeai==0.3.2